src/main/java/chess/domain/Position.java
package chess.domain;

import java.util.Objects;

public class Position {

    private final int row;
    private final int col;

    public Position(final int row, final int col) {
        this.row = row;
        this.col = col;
    }

    public Position move(final UnitVector unitVector) {
        int nextRow = this.row + unitVector.getRow();
        int nextCol = this.col + unitVector.getCol();

        return new Position(nextRow, nextCol);
    }

    public int getRow() {
        return row;
    }

    public int getCol() {
        return col;
    }

    @Override
    public boolean equals(final Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Position position = (Position) o;
        return row == position.row && col == position.col;
    }

    @Override
    public int hashCode() {
        return Objects.hash(row, col);
    }
}
